requirement.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
filelock==3.0.12
gunicorn==20.1.0
Pillow==9.1.1
platformdirs==2.2.0
psycopg2==2.9.3
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
virtualenv==20.7.2
whitenoise==6.2.0
